+++ cmd/hook/hook.go
@@ -16,6 +16,7 @@ func NewHookCommand(client client.API, w io.Writer) *cobra.Command {
 
 	cmd.AddCommand(
 		NewListCommand("list", client, w),
+		NewEventTypeCommand("event-types", client, w),
 	)
 	return cmd
 }